refactor: rename benchmark package
