+++ benchmarks/src/test/java/io/lonmstalker/tgkit/benchmark/TelegramSenderBenchmarkTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.lonmstalker.tgkit.benchmarks;
+package io.lonmstalker.tgkit.benchmark;
 
 import org.junit.jupiter.api.Test;
